requirements.txt
tensorflow-gpu~=2.4.1
numpy~=1.19.5
Pillow~=8.2.0
tqdm~=4.60.0